move order of hosts to query to a Router() class from Transport
